refactor: 특정 기능을 지닌 객체 또는 로직은 libs 폴더 내에서 관린한다.
